chore: add declaration files
